requirements.txt
Django==3.2
PyMySQL==1.0.2
Pillow==8.2.0
